refactor create table
