+++ crates/sail-delta-lake/src/lib.rs
@@ -16,6 +16,6 @@ pub use deltalake::protocol::{DeltaOperation, SaveMode};
 pub use deltalake::DeltaTable;
 pub use operations::{LoadBuilder, SailDeltaOps, WriteBuilder};
 pub use table::{
-    create_delta_table_provider_with_object_store, create_delta_table_with_object_store,
-    open_table_with_object_store,
+    create_delta_provider, create_delta_table_provider_with_object_store,
+    create_delta_table_with_object_store, open_table_with_object_store,
 };
