+++ crates/sail-delta-lake/src/table.rs
@@ -1,11 +1,13 @@
 use std::sync::Arc;
 
+use datafusion::datasource::listing::ListingTableUrl;
+use datafusion_common::Result;
 use deltalake::logstore::{default_logstore, LogStoreRef, StorageConfig};
 use deltalake::{DeltaResult, DeltaTable, DeltaTableError};
 use object_store::ObjectStore;
 use url::Url;
 
-use crate::delta_datafusion::{DeltaScanConfig, DeltaTableProvider};
+use crate::delta_datafusion::{delta_to_datafusion_error, DeltaScanConfig, DeltaTableProvider};
 
 pub async fn open_table_with_object_store(
     table_uri: impl AsRef<str>,
@@ -106,4 +108,53 @@ pub async fn create_delta_table_provider_with_object_store(
 
     // Create the DeltaTableProvider
     DeltaTableProvider::try_new(snapshot, log_store, config)
+}
+
+/// Creates a Delta table provider using SessionContext to get the object store.
+///
+/// This function contains the complete logic for creating a Delta table provider,
+/// including parsing the table URI, getting the object store, and configuring the scan.
+///
+/// # Arguments
+///
+/// * `ctx` - The DataFusion SessionContext
+/// * `table_uri` - The URI of the Delta table
+/// * `options` - Additional options for the Delta table (currently unused)
+///
+/// # Returns
+///
+/// A `Result<Arc<dyn TableProvider>>` that can be used with DataFusion.
+pub async fn create_delta_provider(
+    ctx: &datafusion::prelude::SessionContext,
+    table_uri: &str,
+    options: &std::collections::HashMap<String, String>,
+) -> Result<std::sync::Arc<dyn datafusion::catalog::TableProvider>> {
+    // TODO: Parse options using DataSourceOptionsResolver when needed
+    // let resolver = DataSourceOptionsResolver::new(ctx);
+    // let delta_options = resolver.resolve_delta_read_options(options.clone())?;
+    let _ = options; // Suppress unused variable warning
+
+    let url = ListingTableUrl::parse(table_uri)?;
+    let object_store = ctx.runtime_env().object_store(&url)?;
+
+    let storage_config = StorageConfig::default();
+
+    // Create DeltaScanConfig with proper field names
+    let scan_config = DeltaScanConfig {
+        file_column_name: None,
+        wrap_partition_values: true,
+        enable_parquet_pushdown: true, // Default to true for now
+        schema: None,
+    };
+
+    let table_provider = create_delta_table_provider_with_object_store(
+        table_uri,
+        object_store,
+        storage_config,
+        Some(scan_config),
+    )
+    .await
+    .map_err(delta_to_datafusion_error)?;
+
+    Ok(Arc::new(table_provider))
 }
